+++ lib/ucx_ucc_web/query.ex
@@ -26,11 +26,11 @@ defmodule UcxUccWeb.Query do
   end
 
   def insert(socket, item, opts) do
-    do_update(socket, item, to_map(opts), @no_broadcast)
+    do_insert(socket, item, to_map(opts), @no_broadcast)
   end
 
   def insert!(socket, item, opts) do
-    do_update(socket, item, to_map(opts), @broadcast)
+    do_insert(socket, item, to_map(opts), @broadcast)
   end
 
   def do_insert(socket, type, %{set: set, on: on}, fun) do
@@ -50,20 +50,21 @@ defmodule UcxUccWeb.Query do
   end
 
   defp build_js(selector, {method, attr}, value) do
-    "$('#{selector}').#{method}('#{attr}',#{escape_value(value)});"
+    "$('#{selector}').#{method}('#{attr}',#{escape_value(value)});" # |> IO.inspect(label: "build_js 1")
   end
 
   defp build_js(selector, method, value) when method in ~w(html replaceWith) do
-    "$('#{selector}').#{method}(#{escape_value(value)});"
+    "$('#{selector}').#{method}(#{escape_value(value)});" # |> IO.inspect(label: "build_js 2")
   end
   defp build_js(selector, method, value) do
-    "$('#{selector}').#{method}('#{value}');"
+    "$('#{selector}').#{method}('#{value}');" # |> IO.inspect(label: "build_js 3")
   end
 
   defp to_map(opts), do: Enum.into(opts, %{})
 
   defp jquery_method(:update, :class), do: {"attr", "class"}
   defp jquery_method(:update, :text), do: "text"
+  defp jquery_method(:update, :value), do: "val"
   defp jquery_method(:update, :html), do: "html"
   defp jquery_method(:update, {:class, :toggle}), do: "toggleClass"
   defp jquery_method(:update, :replaceWith), do: "replaceWith"