+++ plugins/ucc_chat/lib/ucc_chat/services/channel_service.ex
@@ -284,6 +284,25 @@ defmodule UccChat.ChannelService do
   end
   def unhide_current_channel(cc, _channel_id), do: cc
 
+  def get_side_nav_search(user, match, channel_id, opts \\ [])
+  def get_side_nav_search(%User{} = user, match, channel_id, opts) do
+    channel = get_channel channel_id
+    rooms = side_nav_search(user, match, channel, opts)
+    %{
+      room_types: rooms |> all_room_types() |> room_types(false),
+      room_map: room_map(rooms),
+      rooms: [],
+      active_room: Enum.find(rooms, &(&1[:open])) || build_active_room(channel),
+      search_empty: rooms == [],
+    }
+  end
+
+  def get_side_nav_search(user_id, match, channel_id, opts) do
+    user_id
+    |> Accounts.get_user
+    |> get_side_nav_search(match, channel_id, opts)
+  end
+
   @doc """
   Get the side_nav data used in the side_nav templates
   """
@@ -335,6 +354,18 @@ defmodule UccChat.ChannelService do
     }
     |> Hooks.build_sidenav_room()
   end
+
+  defp side_nav_search(user, match, channel, opts \\ []) do
+    match
+    |> run_search(user.id, opts[:fuzzy])
+    |> Enum.map(fn cc -> channel_room(cc, user.id, channel, channel.id) end)
+    |> Enum.sort(fn a, b ->
+      String.downcase(a.display_name) < String.downcase(b.display_name)
+    end)
+  end
+
+  defp run_search(match, user_id, true),  do: Subscription.fuzzy_search(match, user_id)
+  defp run_search(match, user_id, _),  do: Subscription.search(match, user_id)
 
   defp side_nav_rooms(user, channel, channel_id, chat_mode) do
     user